+++ src/App.tsx
@@ -1,6 +1,6 @@
 import actionSettingsIcon from '@ui5/webcomponents-icons/dist/action-settings.js';
 import { Avatar, ShellBar } from '@ui5/webcomponents-react';
-import { SavedLocationList } from "./SavedLocationList.tsx";
+import { SavedWeatherList } from "./SavedWeatherList.tsx";
 import {SearchLocation} from "./SearchLocation.tsx";
 import {CurrentWeather} from "./CurrentWeather.tsx";
 
@@ -19,7 +19,7 @@ function App() {
         />
         <SearchLocation/>
         <CurrentWeather/>
-        <SavedLocationList/>
+        <SavedWeatherList/>
     </>
   );
 }
